sensor/control/requirements.txt
RPi.GPIO==0.7.0
paho_mqtt==1.5.0